refactor index.tsx to split out Question, Part, BodyItem and Rubric editing into their own files
